reorganized files and combined all tree-like methods and functions into tree directory
