--- code-challenges/python401/tree-breadth/test_breadth_first.py
@@ -1,14 +1,0 @@
-# from breadth_first import breadth_first
-# from support_items import BinarySearchTree, Queue
-
-
-# def test_breadth_traversal():
-#     tree = BinarySearchTree()
-#     queue = Queue()
-#     tree.add(5)
-#     tree.add(4)
-#     tree.add(10)
-#     tree.add(2)
-
-#     assert breadth_first(tree) == [5, 4, 10, 2]
-#     assert breadth_first(tree) != [5, 4, 10, 3]
